+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [Unreleased]
+- Check the original URL path against the push manifest in addition to the resolved filename. This allows mapping application route patterns to push resources.
 
 ## [0.10.0] 2017-09-11
 - Push manifest keys are now regular expression patterns instead of exact paths.